feat: started etl endpoint for instance
