+++ iaso/api/instances/pagination.py
@@ -1,0 +1,5 @@
+from iaso.api.common import Paginator
+
+
+class ETLInstancePagination(Paginator):
+    page_size = 20
