+++ iaso/api/tasks/create/instance_bulk_gps_push.py
@@ -2,7 +2,7 @@ from drf_spectacular.utils import extend_schema
 from rest_framework import permissions, status, viewsets
 from rest_framework.response import Response
 
-from iaso.api.instances.instances import HasInstanceBulkPermission
+from iaso.api.instances.views import HasInstanceBulkPermission
 from iaso.api.org_units import HasCreateOrgUnitPermission
 from iaso.api.tasks.serializers import TaskSerializer
 from iaso.tasks.instance_bulk_gps_push import instance_bulk_gps_push
