+++ iaso/tasks/instance_bulk_gps_push.py
@@ -9,7 +9,7 @@ from django.http import QueryDict
 
 from beanstalk_worker import task_decorator
 from hat.audit import models as audit_models
-from iaso.api.instances.instance_filters import parse_instance_filters
+from iaso.api.instances.filters import parse_instance_filters
 from iaso.models import Instance, Task
 from iaso.utils.gis import convert_2d_point_to_3d
 from iaso.utils.models.common import check_instance_bulk_gps_push
