+++ iaso/tasks/process_mobile_bulk_upload.py
@@ -28,7 +28,7 @@ from hat.api.export_utils import timestamp_to_utc_datetime
 from hat.api_import.models import APIImport
 from hat.audit.models import BULK_UPLOAD, BULK_UPLOAD_MERGED_ENTITY, log_modification
 from hat.sync.views import create_instance_file, process_instance_file
-from iaso.api.instances.instances import import_data as import_instances
+from iaso.api.instances.views import import_data as import_instances
 from iaso.api.org_unit_change_requests.serializers import OrgUnitChangeRequestWriteSerializer
 from iaso.api.org_units import import_org_units
 from iaso.api.stocks.utils import import_stock_ledger_items
