+++ iaso/urls_etl.py
@@ -1,0 +1,11 @@
+from rest_framework import routers
+
+from iaso.api.instances.views_etl import ETLInstanceViewSet
+
+
+router = routers.DefaultRouter()
+
+router.register(r"instances", ETLInstanceViewSet, basename="instances")
+
+
+urlpatterns = router.urls
